Extract reusable PatchListView component
